+++ commands/ItemCommands.cpp
@@ -6,15 +6,15 @@
 void executeAddItem(const Command& command, GameState& gameState) {
 	gameState.getInventory().addItem(command.id, command.itemQuantity);
 
-	const auto& item = ItemRegistry::instance().get(command.id);
-	std::cout << "Added " << command.itemQuantity << " x " << item.name << " to inventory.\n";
+	auto item = ItemRegistry::instance().get(command.id);
+	std::cout << "Added " << command.itemQuantity << " x " << item->name << " to inventory.\n";
 }
 
 void executeRemoveItem(const Command& command, GameState& gameState) {
 	gameState.getInventory().removeItem(command.id, command.itemQuantity);
 
-	const auto& item = ItemRegistry::instance().get(command.id);
-	std::cout << "Removed " << command.itemQuantity << " x " << item.name << " from inventory.\n";
+	auto item = ItemRegistry::instance().get(command.id);
+	std::cout << "Removed " << command.itemQuantity << " x " << item->name << " from inventory.\n";
 }
 
 void executeUseItem(const Command& command, GameState& gameState) {
@@ -24,9 +24,9 @@ void executeUseItem(const Command& command, GameState& gameState) {
 		return;
 	}
 
-	const auto& item = ItemRegistry::instance().get(command.id);
+	auto item = ItemRegistry::instance().get(command.id);
 
-	if (item.type == ItemType::Key) {
+	if (item->type == ItemType::Key) {
 		std::cout << "You can't use that item here. \n";
 		return;
 	}
@@ -35,12 +35,12 @@ void executeUseItem(const Command& command, GameState& gameState) {
 	Character& target = *party->getMemberByIndex(command.targetIndex);
 	Character& user = target; // for now, assume self-use
 
-	const auto& effect = EffectRegistry::instance().get(item.effectID);
-	int magnitude = item.magnitudeOverride.value_or(command.magnitudeOverride);
+	const auto& effect = EffectRegistry::instance().get(item->effectID);
+	int magnitude = item->magnitudeOverride.value_or(command.magnitudeOverride);
 
 	effect.apply(gameState, user, target, magnitude);
 
 	inventory.removeItem(command.id, command.itemQuantity);
 
-	std::cout << "You used " << item.name << ".\n";
+	std::cout << "You used " << item->name << ".\n";
 }