+++ entities/Character.h
@@ -1,18 +1,27 @@
 #pragma once
 
+#include "equipment/Slot.h"
+#include "equipment/Equippable.h"
 #include <unordered_map>
 #include <string>
+#include <array>
+#include <vector>
+#include <memory>
+
 
 class Character
 {
 public:
-	Character(std::string id, std::string name, int maxHp, bool isMercenary = false);
+	Character(std::string id, std::string name, bool isMercenary = false);
+
+	static constexpr std::size_t SlotCount = static_cast<std::size_t>(Slot::COUNT);
 
 	const std::string& getID() const;
 	const std::string& getName() const;
 	int getLevel() const;
+	int getBaseStat(std::string_view name) const;
 	int getStat(std::string_view name) const;
-	int getSkillBonus(int baseMagnitude, std::string_view stat) const;
+	int getSkillBonus(std::string_view stat) const;
 	int getHP() const;
 	int getMaxHP() const;
 	bool isMercenary() const;
@@ -25,11 +34,21 @@ public:
 
 	bool isAlive() const;
 
+	void rescaleHP(int oldMax);
+
+	std::vector<std::shared_ptr<Equippable>> equip(const std::shared_ptr<Equippable>& item);
+	std::shared_ptr<Equippable> unequip(Slot slot);
+	bool hasEquipped(Slot slot) const;
+	bool isEquipped(std::string_view itemID) const;
+	std::shared_ptr<Equippable> getEquipped(Slot slot) const;                                // returns equipment in slot
+	const std::array<std::shared_ptr<Equippable>, SlotCount>& getEquipment() const noexcept; // returns private array equipment_
+	std::vector < std::shared_ptr<Equippable>> getAllEquipped() const;                      // returns vector of filtered equipment (no duplicates/nullptr)
+
 private:
-	std::string id;
-	std::string name;
-	int level = 1;
-	std::unordered_map<std::string, int> baseStats{
+	std::string id_;
+	std::string name_;
+	int level_ = 1;
+	std::unordered_map<std::string, int> baseStats_{
 		{ "resolve", 5 },
 		{ "power", 5 },
 		{ "tempo", 5 },
@@ -37,8 +56,8 @@ private:
 		{ "clarity", 5 },
 		{ "presence", 5 }
 	};
-	int hp;
-	int maxHP;
-	bool mercenary;
+	int hp_;
+	bool mercenary_;
+	std::array<std::shared_ptr<Equippable>, SlotCount> equipment_;
 };